requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
requests
google-generativeai
logging
